♻️  type 중복 types 폴더로 이동
